+++ src/model/CopyTask.ts
@@ -12,9 +12,9 @@
 export default class CopyTask {
   public projectDepPath: string;
   public modulePath: string;
-  public files: Array<string>;
+  public files: string[];
 
-  constructor(projectDepPath: string, modulePath: string, files: Array<string>) {
+  constructor(projectDepPath: string, modulePath: string, files: string[]) {
     this.projectDepPath = projectDepPath;
     this.modulePath = modulePath;
     this.files = files;