+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "Node Workspace Builder" extension will be documented in this file.
 
+## 0.2.0
+- `Changed` When you run a build task, this extension will update your project dependencies which are not satisfied with version in your `package.json` for now.
+- `Fixed` When you removed some files from your modules, the build task ran into a problem.
+
 ## 0.1.14
 - `Added` Added configuration `node-workspace-builder.autoBuildOnStartUp` to specify if build immediately on start up.